remove some more trash scripts that shouldn't be in the main repo
these are all helper scripts specific to my system that aren't
really useful to anyone.

also made it generally the case that scripts in the root directory
are ignored. "generally useful" tools are put in the `tools` directory.
stuff in the root directory is just notes and bits for me to chew on.