feat(vault): implement basics of dupe detection
Still needs a way to deterministically generate a key for each entry.

Idea: a `hash() -> String` method on StorageContent so that new_record
can call it, store said object with the newly-generated key.

Since this is a deterministic operation, restoring a backup with dupes
would not overwrite what's present in PDDB already.