fix(vault): storage for totp and passwords derive pddb key from their unique content
Instead of generating a unique GUID every time a new record gets
created, hash known unique content for every entry and use that as key.

The StorageContent trait now has a hash() method that returns a Vec
containing the SHA-256 hash of the object's unique content.

Hashing is done by re-using the same calls that ctap uses.

This allows dupes detection when restoring a backup, allowing us to
employ a "never overwrite" policy: Precursor is the main source of truth
for such data.