add "perf" variant to targets, for performance profiling
also add a routine to manually switch over the UTRA in this case.
for whatever reason, the system is not picking up the change in
the UTRA when specified as an environment variable change --
maybe there is a hard coded link somewhere, but I can't seem
to find it.

since performnace counting is an edge case use mostly by
frankly just me, I'm not going to worry about ergonomics so
much.