add persistent default text for issue #281
this adds a method field_placeholder_persist() to the
text entry box builder. When this is used (instead of
field()), the placeholder, if provided, is treated as
"persistent" data, that is, when you press a key, it is
not automatically cleared, but instead, the user starts
editing the existing data.